requirements.txt
anthropic>=0.39.0
pymongo>=4.6.0
voyageai>=0.3.0
requests>=2.31.0
flask>=3.0.0
python-dotenv>=1.0.0
rich>=13.7.0
